人工智能化学分析/week02/requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
pandas>=2.2.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
